python/pyproject.toml
[project]
name = "cora-python"
version = "0.1.0"
description = "Python integration for CORA reachability analysis of neural network controlled systems"
requires-python = ">=3.11"
dependencies = [
    "matlabengine",
    "pyyaml",
    "numpy",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
